requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arch==7.2.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
coverage==7.7.1
cycler==0.12.1
fastapi==0.115.11
fonttools==4.57.0
frozendict==2.4.6
generalized_timeseries==0.1.17
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.1
multitasking==0.0.11
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pandas-datareader==0.10.0
patsy==1.0.1
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1
starlette-graphene3==0.6.0
statsmodels==0.14.4
tabulate==0.9.0
timeseries_compute==0.2.37
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
yfinance==0.2.55
